refactor(headless): extract fixed prompt WAL types
